Extract MethodInvocationCollector to helper
