+++ java-checks/src/main/java/org/sonar/java/checks/tests/AbstractOneExpectedExceptionRule.java
@@ -19,30 +19,24 @@
  */
 package org.sonar.java.checks.tests;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import org.sonar.java.checks.helpers.MethodTreeUtils;
 import org.sonar.java.model.ExpressionUtils;
 import org.sonar.plugins.java.api.IssuableSubscriptionVisitor;
 import org.sonar.plugins.java.api.JavaFileScannerContext;
 import org.sonar.plugins.java.api.semantic.MethodMatchers;
-import org.sonar.plugins.java.api.semantic.Symbol;
 import org.sonar.plugins.java.api.semantic.Type;
 import org.sonar.plugins.java.api.tree.Arguments;
-import org.sonar.plugins.java.api.tree.BaseTreeVisitor;
-import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.ExpressionStatementTree;
 import org.sonar.plugins.java.api.tree.ExpressionTree;
 import org.sonar.plugins.java.api.tree.IdentifierTree;
 import org.sonar.plugins.java.api.tree.LambdaExpressionTree;
 import org.sonar.plugins.java.api.tree.MemberSelectExpressionTree;
 import org.sonar.plugins.java.api.tree.MethodInvocationTree;
-import org.sonar.plugins.java.api.tree.NewClassTree;
 import org.sonar.plugins.java.api.tree.StatementTree;
 import org.sonar.plugins.java.api.tree.Tree;
 import org.sonar.plugins.java.api.tree.TryStatementTree;
@@ -193,41 +187,6 @@ public abstract class AbstractOneExpectedExceptionRule extends IssuableSubscript
     return methodInvocationTrees.stream()
       .map(expr -> new JavaFileScannerContext.Location(message, expr))
       .collect(Collectors.toList());
-  }
-
-  static class MethodInvocationCollector extends BaseTreeVisitor {
-    List<Tree> invocationTree = new ArrayList<>();
-    private final Predicate<Symbol> collectPredicate;
-
-    MethodInvocationCollector(Predicate<Symbol> collectPredicate) {
-      this.collectPredicate = collectPredicate;
-    }
-
-    @Override
-    public void visitMethodInvocation(MethodInvocationTree mit) {
-      if (collectPredicate.test(mit.symbol())) {
-        invocationTree.add(ExpressionUtils.methodName(mit));
-      }
-      super.visitMethodInvocation(mit);
-    }
-
-    @Override
-    public void visitNewClass(NewClassTree tree) {
-      if (collectPredicate.test(tree.constructorSymbol())) {
-        invocationTree.add(tree.identifier());
-      }
-      super.visitNewClass(tree);
-    }
-
-    @Override
-    public void visitClass(ClassTree tree) {
-      // Skip class
-    }
-
-    @Override
-    public void visitLambdaExpression(LambdaExpressionTree lambdaExpressionTree) {
-      // Skip lambdas
-    }
   }
 
 }
